requirements.txt
requests
aiohttp<3.8.0,>=3.6.0
async-timeout<4.0,>=3.0
certifi==2018.11.29
chardet==3.0.4
discord.py==1.7.3
multidict==4.5.2
websockets==9.1
ojichat==1.0.0
